Moved AmazonSqsConsumerBase and its interfaces to Common project
