+++ NationalArchives.Taxonomy.Batch/appsettings.json
@@ -123,5 +123,5 @@
     "CollectionName": "categories"
   },
   "CategorySource": "Mongo",
-  "OperationMode": "Full_Reindex" // "Daily_Update" or "Full_Reindex"
+  "OperationMode": "Daily_Update" // "Daily_Update" or "Full_Reindex"
 }
