+++ Taxonomy.Common/Domain/Queue/ISourceIaidInputQueueConsumer.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 
 namespace NationalArchives.Taxonomy.Batch.DailyUpdate.MesssageQueue
 {
-    interface ISourceIaidInputQueueConsumer : IDisposable
+    public interface ISourceIaidInputQueueConsumer : IDisposable
     {
 
         Task Init(CancellationToken token);
