+++ Taxonomy.Common/Domain/Queue/ISourceIaidInputQueueConsumerAdapter.cs
